Track watered status of a plant
This commit adds a way to track the watered status of a plant based on
how frequently a plant likes to be watered.

Given a wateringFrequencyInDays attribute on a plant, we can determine
if a plant is overwatered, underwatered or healthy based on how many
times it was watered within that time frame. For example, with  a plant
with a wateringFrequencyInDays of 3:

A plant with one recorded watering in the past 3 days is "Healthy"
A plant with no recorded waterings is "Underwatered"
A plant with more than one watering is "Overwatered"

Next step is probably implementing something to apply a penalty to a
plant's health based on how long a plant is considering over- or
underwatered.